changed indice type from float to long long

+++ nn/prediction_batch_provider.hpp
@@ -10,7 +10,7 @@ class PredictionBatchProvider {
  public:
   PredictionBatchProvider(std::vector<float> sample_data, DataShape sample_dim);
   PredictionBatchProvider(af::array samples);
-  PredictionBatch batch(std::vector<float> indices);
+  PredictionBatch batch(std::vector<long long> indices);
   PredictionBatch batch(unsigned int from, unsigned int batch_size);
   unsigned int sample_count();
   af::dim4 sample_dimensions();
