+++ nn/training_batch_provider.hpp
@@ -12,7 +12,7 @@ class TrainingBatchProvider {
   TrainingBatchProvider(std::vector<float> sample_data, DataShape sample_dim,
                         std::vector<float> target_data, DataShape target_dim);
   TrainingBatchProvider(af::array samples, af::array targets);
-  TrainingBatch batch(std::vector<float> indices);
+  TrainingBatch batch(std::vector<long long> indices);
   TrainingBatch batch(unsigned int from, unsigned int batch_size);
   unsigned int sample_count();
   af::dim4 sample_dimensions();
